Refactor : rename some variables and method to make the logic more clear
