+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-integration-test-scaling/src/test/java/org/apache/shardingsphere/integration/data/pipeline/framework/container/compose/DockerComposedContainer.java
@@ -19,7 +19,7 @@ package org.apache.shardingsphere.integration.data.pipeline.framework.container.
 
 import lombok.Getter;
 import org.apache.shardingsphere.infra.database.type.DatabaseType;
-import org.apache.shardingsphere.integration.data.pipeline.framework.container.config.proxy.ScalingProxyClusterContainerConfiguration;
+import org.apache.shardingsphere.integration.data.pipeline.framework.container.config.proxy.ScalingProxyClusterContainerConfig;
 import org.apache.shardingsphere.integration.data.pipeline.framework.container.config.storage.ScalingStorageContainerConfigurationFactory;
 import org.apache.shardingsphere.test.integration.env.container.atomic.adapter.AdapterContainerFactory;
 import org.apache.shardingsphere.test.integration.env.container.atomic.adapter.impl.ShardingSphereProxyClusterContainer;
@@ -49,7 +49,7 @@ public final class DockerComposedContainer extends BaseComposedContainer {
         governanceContainer = getContainers().registerContainer(new ZookeeperContainer());
         storageContainer = getContainers().registerContainer((DockerStorageContainer) StorageContainerFactory.newInstance(databaseType, dockerImageName,
                 "", ScalingStorageContainerConfigurationFactory.newInstance(databaseType, "")));
-        ScalingProxyClusterContainerConfiguration scalingProxyClusterContainerConfiguration = new ScalingProxyClusterContainerConfiguration(databaseType, dockerImageName);
+        ScalingProxyClusterContainerConfig scalingProxyClusterContainerConfiguration = new ScalingProxyClusterContainerConfig(databaseType, dockerImageName);
         ShardingSphereProxyClusterContainer proxyClusterContainer =
                 (ShardingSphereProxyClusterContainer) AdapterContainerFactory.newInstance("Cluster", "proxy", databaseType, storageContainer, "", scalingProxyClusterContainerConfiguration);
         proxyClusterContainer.dependsOn(governanceContainer, storageContainer);
