+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-integration-test-suite/src/test/java/org/apache/shardingsphere/test/integration/container/compose/mode/ClusterComposedContainer.java
@@ -19,7 +19,7 @@ package org.apache.shardingsphere.test.integration.container.compose.mode;
 
 import org.apache.shardingsphere.test.integration.container.compose.ComposedContainer;
 import org.apache.shardingsphere.test.integration.container.config.SuiteStorageContainerConfigurationFactory;
-import org.apache.shardingsphere.test.integration.container.config.proxy.SuiteProxyClusterContainerConfiguration;
+import org.apache.shardingsphere.test.integration.container.config.proxy.SuiteProxyClusterContainerConfig;
 import org.apache.shardingsphere.test.integration.env.container.atomic.DockerITContainer;
 import org.apache.shardingsphere.test.integration.env.container.atomic.ITContainers;
 import org.apache.shardingsphere.test.integration.env.container.atomic.adapter.AdapterContainer;
@@ -54,7 +54,7 @@ public final class ClusterComposedContainer implements ComposedContainer {
         // TODO add more version of databases
         storageContainer = containers.registerContainer(StorageContainerFactory.newInstance(parameterizedArray.getDatabaseType(), "", scenario,
                 SuiteStorageContainerConfigurationFactory.newInstance(parameterizedArray.getDatabaseType(), scenario)));
-        SuiteProxyClusterContainerConfiguration suiteProxyClusterContainerConfiguration = new SuiteProxyClusterContainerConfiguration(scenario);
+        SuiteProxyClusterContainerConfig suiteProxyClusterContainerConfiguration = new SuiteProxyClusterContainerConfig(scenario);
         AdapterContainer adapterContainer = AdapterContainerFactory.newInstance(
                 parameterizedArray.getMode(), parameterizedArray.getAdapter(), parameterizedArray.getDatabaseType(), storageContainer, scenario, suiteProxyClusterContainerConfiguration);
         if (adapterContainer instanceof DockerITContainer) {
